refactor: logout header refactored
