+++ src/components/UI/organism/LogoutHeader/LogoutHeader.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { LogoutHeader } from './LogoutHeader.component'
 
 const meta: Meta<typeof LogoutHeader> = {
-  title: 'Organism/LogoutHeader',
+  title: 'Organism/Logout Header',
   component: LogoutHeader,
   args: {}
 }
